pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "study-evaluation-pipeline"
version = "0.2.0"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv==1.1.1",
    "openai==2.5.0",
    "google-genai==1.45.0",
    "pypdf==6.1.2",
    "tabulate==0.9.0",
    "pyyaml==6.0.3",
    "pandas==2.3.3",
    "pandas-stubs==2.3.2.250926",
    "scikit-learn==1.7.2",
    "typer==0.20.0",
    "fastapi==0.120.1",
    "uvicorn==0.38.0"
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
sep-run = "sep.cli.cli:app"
sep-api-test = "sep.api_request.api_test:main"

sep-create-csv = "sep.evaluation.create_csv:main"
sep-evaluate = "sep.core.evaluation.evaluate_run:main"
sep-display = "sep.display_runs:main"
sep-prompt-designer = "sep.core.prompt_designer.run_prompt_design:main"
